+++ blueprint/implementations/billing/stripe/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @forklaunch/implementation-billing-stripe
 
+## 1.1.19
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/implementation-billing-base@1.0.19
+  - @forklaunch/interfaces-billing@1.0.18
+
+## 1.1.18
+
+### Patch Changes
+
+- update package versions
+- Updated dependencies
+  - @forklaunch/implementation-billing-base@1.0.18
+  - @forklaunch/interfaces-billing@1.0.17
+
+## 1.1.17
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/implementation-billing-base@1.0.17
+  - @forklaunch/interfaces-billing@1.0.16
+
 ## 1.1.16
 
 ### Patch Changes