+++ blueprint/implementations/worker/database/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @forklaunch/implementation-worker-database
 
+## 1.0.20
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.18
+
+## 1.0.19
+
+### Patch Changes
+
+- update package versions
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.17
+
+## 1.0.18
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.16
+
 ## 1.0.17
 
 ### Patch Changes