+++ blueprint/interfaces/worker/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @forklaunch/interfaces-worker
 
+## 1.0.18
+
+### Patch Changes
+
+- update packages
+
+## 1.0.17
+
+### Patch Changes
+
+- update package versions
+
+## 1.0.16
+
+### Patch Changes
+
+- update packages
+
 ## 1.0.15
 
 ### Patch Changes